% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandatapkgs.R
\name{available_data}
\alias{available_data}
\title{available_data}
\usage{
available_data(fields = c("Package", "Version", "License", "Depends",
  "Imports", "Suggests", "Enhances"), ...)
}
\arguments{
\item{fields}{A character vector of fields to extract from \code{\link[utils]{available.packages}}.}

\item{\dots}{Additional arguments passed to \code{\link[utils]{available.packages}}.}
}
\value{
A data.frame containing the requested fields, plus \dQuote{Type} (indicating whether it is a data-only package or a data-supplement to another package) and \dQuote{DataDescription} (providing a short description of the data contained in the package).
}
\description{
Find Data-only Packages on CRAN
}
\details{
This function returns a data.frame representation of the output of \code{\link[utils]{available.packages}} that contains only known data-only and/or data-heavy packages that are available on CRAN.
}
\examples{
\dontrun{
# get available data packages
a <- available_data()
}
}

