% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandatapkgs.R
\docType{package}
\name{crandatapkgs-package}
\alias{crandatapkgs}
\alias{crandatapkgs-package}
\title{Find Data-only Packages on CRAN}
\description{
Find Data-only Packages on CRAN
}
\details{
This package provides a subset of \code{available.packages()}
}
\seealso{
\code{\link{available_data}}
}

